third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/paint/hit-testing-001.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: hit testing</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<link rel="match" href="scroll-into-view-ref.html">
<script src="../resources/utils.js"></script>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
body {
  margin: 0;
  padding: 0;
}
#outer {
  width: 100px;
  height: 100px;
  background: lightblue;
}
#inner {
  margin: 25px;
  width: 50px;
  height: 50px;
  background lightgreen;
}
</style>

<body id="body">
<div id="outer"><div id="inner"></div></div>
</body>

<script>
async_test((t) => {
  async function runTest() {
    const container = document.getElementById("outer");
    await setInvisible(container);
    let target = document.elementFromPoint(50, 50);
    t.step(() => assert_equals(target.id, "outer", "center hits outer"));
    target = document.elementFromPoint(10, 50);
    t.step(() => assert_equals(target.id, "outer", "edge hits outer"));
    target = document.elementFromPoint(100, 50);
    t.step(() => assert_equals(target.id, "body", "elsewhere hits body"));
    t.done();
  }
  window.onload = () => { requestAnimationFrame(runTest); };
});

</script>
</html>
